requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-mysql==4.0.0
Faker==8.12.1
idna==2.10
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
